requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
google-generativeai
